feat: make adapter -> enterprise communication async
Makes all API calls to Enterprise asyncronous by using the result store
to maintain state in memory of requests.

The exception is the call to the feeds endpoint as that is used to
verify the connection between the adapter and Enterprise so that should
continue to be a synchronous request.

Signed-off-by: Bradley Jones <bradley.jones@anchore.com>